requirements.txt
biopython>=1.79
joblib>=1.2.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
streamlit>=1.30.0
